Run at document-start
Awesome userscript, I love it.  But it would often flicker for me, only making the content wide *after* it was already being displayed.

I tried adding `@run-at document-start` a few months ago, which solved the problem, and so far has not caused any issues.  (TamperMonkey in Chrome, Greasemonkey in Firefox, Linux and Mac.)