Fix to SPDX
OUJS has made a change recently to require SPDX codes for OSI approved licensing.

In order to improve the appearance of your script homepages it would be appreciated if you could modify this affected script.

Until this change is made you will be unable to update this script.

Thanks for your support.
OUJS Staff